+++ fizzbuzz/discord/extensions/admin/__init__.py
@@ -1,4 +1,4 @@
-"""All of the extensions (Cogs) for FizzBuzz."""
+"""All of the admin extensions for FizzBuzz."""
 
 # this file intentionally left empty.
 # python simply requires a __init__ file for importing.