+++ fizzbuzz/discord/extensions/public/__init__.py
@@ -1,4 +1,4 @@
-"""All of the extensions (Cogs) for FizzBuzz."""
+"""All of the global extensions for FizzBuzz."""
 
 # this file intentionally left empty.
 # python simply requires a __init__ file for importing.